feat: separating cryptography service into signing and encryption services
